updated task2

+++ Assignment_3/A3T2/task2.py
@@ -1,5 +1,5 @@
 from pyspark.sql import SparkSession
-from pyspark.sql.types import DoubleType
+from pyspark.sql.types import FloatType
 from pyspark.sql.functions import countDistinct
 from pyspark import SparkContext
 from pyspark.sql import SQLContext
@@ -13,8 +13,8 @@ spark = SparkSession.builder.appName("A3T2").getOrCreate()
 city_df = spark.read.option("header",True).csv(sys.argv[1])
 global_df = spark.read.option("header",True).csv(sys.argv[2])
 
-city_df= city_df.withColumn("AverageTemperature", city_df["AverageTemperature"].cast(DoubleType()))
-global_df= global_df.withColumn("LandAverageTemperature", global_df["LandAverageTemperature"].cast(DoubleType()))
+city_df= city_df.withColumn("AverageTemperature", city_df["AverageTemperature"].cast(FloatType()))
+global_df= global_df.withColumn("LandAverageTemperature", global_df["LandAverageTemperature"].cast(FloatType()))
 
 cond= [city_df["dt"]==global_df["dt"], city_df["AverageTemperature"] > global_df["LandAverageTemperature"]]
 res= city_df.join(global_df, cond, "inner").select(city_df.Country, city_df.dt)
